requirements.txt
# uDMXController
pyusb
udmx-pyusb

# OpenDMXController
pyftdi

# Serial Controller
pyserial

# Web
flask